rf: move `JsonValidationErrorView`
From `cli/models.py` to `models.py`
